Added an option to aim only while shooting.
The user will be able to select if they want the aimbot to be enabled only while they are shooting or be enabled all the time (or on key down). I personally like to have this option disabled so I can hold my aim hotkey to help me aim before I shoot.