Optimize entity classification performance
Classify entities using cs2::CEntityClass* from cs2::CEntityIdentity
instead of the VMT pointer of the entity object. Due to the structure of
the entity list, classification using entity class pointer is ~2.4x faster
(as the entity identity is already in the CPU cache) than using the VMT pointer
which requires a fetch from the main memory.